+++ mayan/apps/document_states/tests/test_workflow_actions.py
@@ -7,7 +7,7 @@ from mayan.apps.testing.tests.mixins import TestServerTestCaseMixin
 from mayan.apps.testing.tests.mocks import request_method_factory
 
 from ..literals import WORKFLOW_ACTION_ON_ENTRY
-from ..models import Workflow
+from ..models import Workflow, WorkflowInstance
 from ..permissions import permission_workflow_template_edit
 from ..workflow_actions import (
     DocumentPropertiesEditAction, DocumentWorkflowLaunchAction, HTTPAction
@@ -301,7 +301,9 @@ class DocumentPropertiesEditActionTestCase(
             action_path=TEST_DOCUMENT_EDIT_WORKFLOW_TEMPLATE_STATE_ACTION_DOTTED_PATH,
             label='', when=WORKFLOW_ACTION_ON_ENTRY,
         )
-        self.test_workflow_template.document_types.add(self.test_document_type)
+        self.test_workflow_template.document_types.add(
+            self.test_document_type
+        )
 
         self._create_test_document_stub()
 
@@ -342,7 +344,7 @@ class DocumentWorkflowLaunchActionTestCase(
         action.execute(context={'document': self.test_document})
 
         self.assertTrue(
-            workflow_count + 1, self.test_document.workflows.count()
+            self.test_document.workflows.count(), workflow_count + 1
         )
 
 
@@ -359,7 +361,8 @@ class DocumentWorkflowLaunchActionViewTestCase(
         self._create_test_workflow_template_state()
 
         self.grant_access(
-            obj=self.test_workflow_template, permission=permission_workflow_template_edit
+            obj=self.test_workflow_template,
+            permission=permission_workflow_template_edit
         )
 
         action_count = self.test_workflow_template_state.actions.count()
@@ -368,5 +371,38 @@ class DocumentWorkflowLaunchActionViewTestCase(
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(
-            self.test_workflow_template_state.actions.count(), action_count + 1
+            self.test_workflow_template_state.actions.count(),
+            action_count + 1
+        )
+
+    def test_document_workflow_launch_action_view_and_document_create_with_full_access(self):
+        self._create_test_workflow_template(
+            add_test_document_type=True, auto_launch=False
+        )
+
+        self._create_test_workflow_template(
+            add_test_document_type=True, auto_launch=True
+        )
+        self._create_test_workflow_template_state()
+
+        self.grant_access(
+            obj=self.test_workflow_template,
+            permission=permission_workflow_template_edit
+        )
+
+        action_count = self.test_workflow_template_state.actions.count()
+        workflow_instance_count = WorkflowInstance.objects.count()
+
+        response = self._request_document_workflow_template_launch_action_create_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.assertEqual(
+            self.test_workflow_template_state.actions.count(),
+            action_count + 1
+        )
+
+        self._create_test_document_stub()
+
+        self.assertEqual(
+            WorkflowInstance.objects.count(), workflow_instance_count + 2
         )